pyproject.toml
[project]
name = "ai-readme-updaterdatasense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0",
    "pygithub>=1.58",
    "gitpython>=3.1",
    "fastapi>=0.95",
    "uvicorn[standard]>=0.22",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.0.25",
    "openai>=1.0.0",
    "pytest",
    "requests>=2.32.5",
    "redis>=6.4.0",
    "rq>=2.6.0",
    "tenacity>=9.1.2",
    "structlog>=25.4.0",
    "prometheus-client>=0.23.1",
    "logging>=0.4.9.6",
]
